feat(core): Create ONNX flavors
