+++ core/src/pipeless_ai/lib/worker/inference/runtime.py
@@ -4,7 +4,31 @@ import onnx
 import onnxruntime
 
 from pipeless_ai.lib.logger import logger
-from pipeless_ai.lib.worker.inference.utils import get_model_path, load_model, parse_input_shape, prepare_frame
+from pipeless_ai.lib.worker.inference.utils import get_model_path, parse_input_shape, prepare_frame
+
+def load_model(file: str, alias: str, force_opset_version: int | None = None, force_ir_version: int | None = None):
+    """
+    Loads a model with onnx and checks it.
+    Returns the model if correct. Finishes execution otherwise.
+    """
+    # TODO: convert the model on the fly checking the extensions from the most common frameworks. Or implement conversion in the CLI instead.
+    try:
+        logger.info(f'Checking {alias} inference model')
+        model = onnx.load(file)
+        if force_opset_version:
+            logger.info(f'Converting from OpSet version {model.opset_import[0].version} to {force_opset_version}')
+            model = onnx.version_converter.convert_version(model, force_opset_version)
+        if force_ir_version:
+            logger.info(f'Converting from IR version {model.ir_version} to {force_opset_version}')
+            model.ir_version = force_ir_version
+        onnx.checker.check_model(model)
+        logger.info(f'Model operation set (OpSet) version: {model.opset_import[0].version}')
+        logger.info(f'Model intermediate representation (IR) version: {model.ir_version}')
+    except Exception as e:
+        logger.error(f'Error loading the {alias} model: {e}')
+        sys.exit(1)
+
+    return model
 
 class PipelessInferenceSession():
     """
@@ -106,13 +130,3 @@ class PipelessInferenceSession():
         except Exception as e:
             logger.error(f'There was an error running inference: {e}')
             return None
-
-def get_inference_session(config):
-    """
-    Returns an inference session when possible or None
-    """
-    if config.get_worker().get_inference().get_model_uri():
-        inference_config = config.get_worker().get_inference()
-        return PipelessInferenceSession(inference_config)
-    else:
-        return None
